bfd: Zero initialise dynamic section contents.
For ARC the dynamic sections are created, and sized early on in the
linking process, within elf_arc_check_relocs, which is called before
garbage collection has taken place.

During elf_arc_check_relocs the dynamic sections are created to hold the
relocations that will be placed in the final executable (or shared
library).  We create a single dynamic section of each name, so there
might be many input sections called .text, all of which require a
dynamic section called .rela.text.  We actually create a single section
called .rela.text that is big enough to hold all of the relocations that
will be placed within it.

The actual allocation of memory for the .rela.text section is done later
within the linking process in elf_arc_size_dynamic_sections, which is
called after linker garbage collection has taken place.

Now it has always been the case that some relocation, that early on we
thought would be copied into the dynamic section, turn out to not be
copied.  We already have a process in place to handle this.

However, garbage collection introduces another issue, rather than
single relocations being dropped from the dynamic section, we remove all
relocations from a particular section.

The two cases above both deal with reducing the size of a dynamic
section based on other events that occur in the linking process, the
first case is already handled while the second is not.

It is tempting to think that the solution to the second problem is easy,
track the number of relocations per section that are to be copied into
the corresponding dynamic section, then when a section is garbage
collected, the dynamic section can have its size reduced.

The problem is that there is a risk of double counting, a relocation might
fall into both counts, and we must not reduce the size of the dynamic
section twice for a single relocation.  To avoid this an additional
layer of complexity would need to be introduced to manage tracking which
relocations are due to be copied to the dynamic section.

This extra complexity is, I believe, the correct, long term solution.
However, I think this complexity should be introduced as part of a
rewrite of how ARC handles creation of dynamic sections, a rewrite that
would move ARC into line with how targets like arm, and x86-64 create
their dynamic sections.

[ The difference is that for arm, and x86-64 (and others) the dynamic
  sections are not sized early, instead information is gathered about
  which relocations will be copied into dynamic sections, this data is
  kept up to date through out the linking process, then finally, the
  dynamic sections are created, and sized correctly. ]

For now, I have solved the problem of garbage collected sections using a
much easier hack.  The dynamic sections are created zero initialised.
Now, relocations are not copied as the source section is garbage
collected then the dynamic section will be left with content set to
zero, which is harmless.

bfd/ChangeLog:

	* elf32-arc.c (elf_arc_size_dynamic_sections): Zero initialise
	dynamic section content.