bfd/arc: Rework relocation alignment decision.
For some relocations the choice of whether to perform the alignment or
not is based on whether the relocation is in a code section.  This
commit adds a new method that returns the required byte-alignment for a
relocation of a given type in a given section, this is then applied to
the relocation value as part of the pc-relative adjustment.

There is one slight issue with this commit.  Some of the relocations,
that were clearly being treated as pc-relative before, are not actually
marked as pc-relative.  There's a cryptic comment in the relocation
table indicating that at some point in time at least one of these
relocations (R_ARC_GOTPC32) was marked as pc-relative.  However it no
longer is, and the comment does not indicate why not.

Given a pressing release deadline I am not currently changing any of the
offending relocations (see arc_relocation_byte_alignment for the list)
to be pc-relative, instead when fixing up pc-relative relocations, I
also treat as pc-relative all relocations that require a byte alignment
greater than 1.  This matches the behaviour that was present before.

At some point after the release the offending relocations should be made
pc-relative, and the extra check of the byte alignment being greater
than 1 should be removed.

There's a new linker test in here that highlights the original issue
that led to this commit.

bfd/ChangeLog:

	* elf32-arc.c (arc_relocation_byte_alignment): New function.
	(elf_arc_relocate_section): Use new function.

ld/testsuite/ChangeLog:

	* ld-arc/unaligned-pc32.d: New file.
	* ld-arc/unaligned-pc32.s: New file.
	* ld-arc/arc.exp: Run new test.

(cherry picked from commit e7bef685f4312b964f6df3cafe9f02e3d3ce4998)